added compression for ActorDatagram and updated Docu for 1.1.1 use compressed datagrams for transmitted state structures

+++ source/nl/plugins/multiplayersample/network/nlDynamicActorReplicaComponent.h
@@ -62,7 +62,7 @@ namespace nl	{
 		SLSize _tickSerialized;
 		bool _sampleVehicleState;
 
-		unsigned int _killCount;
+		unsigned short _killCount;
 
 	private:
 	};
